+++ .phpstorm.meta.php/options.meta.php
@@ -11,7 +11,8 @@ namespace PHPSTORM_META {
 
     registerArgumentsSet(
         'option_names',
-        'docker.client.path', 'docker.socket.path', 'script.exit-early', 'step.clone-path'
+        'docker.client.path', 'docker.socket.path', 'script.bash-runner', 'script.exit-early', 'script.runner',
+        'step.clone-path'
     );
 
     expectedArguments(