+++ src/Utility/Options.php
@@ -29,6 +29,8 @@ class Options
         $definition = array(
             'docker.socket.path' => array('/var/run/docker.sock', null),
             'docker.client.path' => array('/usr/bin/docker', null),
+            'script.runner' => array('/bin/sh', null),
+            'script.bash-runner' => array(true, null),
             'script.exit-early' => array(false, null),
             'step.clone-path' => array('/app', Types::ABSPATH),
         );